backend_/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
psycopg2-binary
pg8000
cloud-sql-python-connector[pg8000]
google-cloud-storage
google-cloud-aiplatform
google-cloud-speech
google-generativeai
vertexai
protobuf
moviepy
reportlab
PyPDF2
python-multipart
pydub
python-jose[cryptography]
